Add D3.js dependency, stash work-in-progress on derived ... ... application build. This commit specifically experiments with creating links between middleforks/node_modules and generated middleforks/DIST/LIB targets.
Ultimately, we want to create two webpack bundles per application
build. One is the Node.js server application (and all its runtime
dependencies in a single .js file). The other is the browser client
application (and all its runtime dependencies in a single .js file).

It is during the creation of an application's webpack bundles
that we actually need to resolve dependencies. The intent is that
the outer monorepo serve as a single point of control over external
dependencies. Currently, we're going to maintain this by-hand.

So, how will a derived application resolve the _latest_ build of
Encapsule Project libraries AND pickup other key dependencies
(e.g. React)?
- Application source code *does not* define a `package.json`.
- Application source code *presumes* that standard references to external packages will resolve correctly.
- Implicitly, external package references in application sources are resolved to the monorepo's `node_modules/*` directory.
    - The set of development dependencies defined in the monorepo package.json is the union of:
        - Monorepo build system dependencies (e.g. babel, webpack)
        - Encapsule Project library dependencies
        - Application dependencies
- Encapsule Project arccore and arctools packages change infrequently and are installed from npmjs.org, and updated manually as required.
- Encapsule Project web application libraries derived from arccore are build and published as part of the monorepo build.
    - Makefile has been modified to create/remove soft links between monorepo/DISTS/LIB/* and monorepo/node_module/*
    - This allows derived applications to import the _latest_ build of these libraries using _standard require/import_ syntax.

b