WIP: Okay... making some progress now. This commit updates the transition... ... operator factory to update the runtime calling signature of the generated filter object to align with the evaluation context names/labels/conventions of the new ObservableProcessController::_evaluate method.
With this, I then updated several existing transition operator filters from RUX
to wrap them in the new TransitionOperator ES6 class. Starting with the logical
operators as they're relatively simple. These are now exported as an "operator set"
which is an array of TransitionOperator class instances.

And, finally with these changes to TransitionOperator we updated the implementation
of ObservableProcessController::_evaluate method so that it dispatches live
transition operator filters!

The existing transition operator filters from RUX project need to get ported over.
But, now I am focused on ObversableProcessController::_evaluate and am moving on
to apply much the same pattern to controller action filters.