WIP: Cleaning up ObservableProcessController::_evaluate method... ... The short-term goal is to get the _evaluate method working completely. The inner loop, the so-called evaluation frame loop appears to be working to a 1st approximation. However, a single call to _evaluate will always execute at least one but typically more separate evaluation frames in sequence until all OPM instances are stable (i.e. evaluation reveals that they do not need to transition between process steps).
So, to complete the bridge from the inner frame loop to
the outer evalution loop means that I need to keep track
of the number of OPM instance step transitions occur
within the frame evaluation.

However, the frame evaluation loop generates a _lot_
of telemetry that's intended to be used as input to development
tools, tests, diagnostics etc. And, I played pretty fast and
loose with the algorithm that generates this data to get
it working. The thought of adding even one or two more
counters, errors, metrics, who knows without getting the
details under control freaked me out. So, now
ObservableController::_evaluate delegates everything
to a filter that uses a reference to the calling class
instance to retain effectively the original calling
semantics of the algorithm. But, now the mass of telemetry
that's produced by a call through OPC::_evaluate is at
last partially schematized using a filter spec. I've already
found a few bugs, changed the schema and used filter to
help me port the implementation...

Taking a snapshot and will pick this up after sleep.