WIP: Stash another huge push in OPC... ... I am adding another plane to the OCP runtime stack:
Firstly, the shared data store is in the process of being
renamed ObserverableControllerData (OCD) because why not.

When you construct an OPC via operator new, you provide
a filter spec for the OPC's contained OCD. This filter
spec defines the layout of the OPC's shared memory.
And, it defines (via __`__appdsl: { opm: "IRUT" };`
annotation) OPM bindings.

For developers this means a large degree of flexibility
because the OCP will find all the OPM instances at in
the OCD at runtime automatically.

Great. The problem:

It's difficult to write the developer-provider OCD spec.

At least, it's been hard to define the developer-provided
OCD spec.

Previously:

YOU: define everything that the alien spaceship needs even though
you really have no fucking idea what's going on. And, you also
define all your own shit. Typically, there are operator and action
filters performing operations on shared OCD data bound to a specific
OPM instance (OPMI). The requirements of these filters in terms
of what specific namespaces they need in your OPM-bound namespace
are, frankly, not something that devleopers should be forced to
maintain in their master OCD filter spec.

So... drum roll.

`ObservableControllerModel` constructor now accepts an optional property `observableControllerDataSpec` that is used inside of its contructor function to **syntesize** the actual OCD store spec that's used at runtime. It is formed via a prioritized merge of OPC namespace requirements, OPM-specific requirements, and your specific requirements as defined in your own OPC spec.

This has the following implications:

- Much simpler for developers to author OCD spec's.
    - Effectively it's a binding frame --- the OCP creates the actual runtime filter spec.
    - OPM declarations (i.e. new ObservableProcessModel) now represent:
        - A template for a descriptor object in the OCD.
        - An FSM model of an observable process that is evaluated dynamically at runtime to invoke operator and action filters that depend on specific namespace name and entity type conventions defined in the OPM's OCD template spec.

So for example we could define an OPM and call it Button Controller. It contains a filter spec template for a descriptor object that defines all the shared data needed to support a Button Controller instance at runtime.

To use our new Button Controller OPM, we create a descriptor object definition in the OCD spec we pass to OPC and bind it via `____appdsl` annotation:

```
const ocdSpec = {
    ____label: "Button Controllers",
    ____description: "
    ____types: "jsArray",
    ____defaultValue: [],
    buttons: {
        ____types: "jsObject",
        ____appdsl: { opm: "f9FbDh5xTJWEaxSD_XqBag" }
    }
};
```