WIP: Bug fixes related to renaming controller data store to... ... observable controller data. This patch updates the evaluation algorithm to use the new location of the ocdi (an ocd class instance) in this._private.ocdi. And, also updates the transition operator and controller action plug-in filter factories that provide implementation for ControllerAction and TransitionOperator ES6 classes.
This commit also includes a small (but stable and growing)
suite a unit tests for OPC implemented using Mocha (seems to be
working very well for this style of tesing that I defined for
onm, and have subsequently refined for filter, them discriminator
and then back to graph with great success. It works --- code
developed in this manner is rock solid. So, over next days
the OPC will rise up and be able to withstand production
workloads and behave itself predictably. At least within the
limits of the test coverage I'm able to get in place.

The goal is to get unit tests on all behaviors that we will
require to work reliably to take a major dependency in derived
vp5 (and later Encapsule.io will be completely rewritten to
use all the latest and greatest -- it's badly dated at this
point. But, whatever for now).