WIP: Experimenting with automated comparison via synthesized `git diff` sub-process calls inside the holodeck test runner. This allows me to use the current branch as the baseline and perist git diff summary markers for eval log files that have been altered by the most-recent execution of the repo's test vectors. What I am looking at is if I can deduce a stable change marker over volatile values in the actual eval log diff. For example, we have seen in OPC already that there's evaluation timing information recorded by the class instance that is reported when you serialize the instance to JSON. These timing marks are not stable (e.g. the actual start/end times are used and these will presumably always be different test-run-to-test-run).
What I think is possible is to deduce a stable git diff
signature that indicates, effectively, we expact the
test run to mutate these regions of data in the eval logs.
And, then if such a diff marker is committed to the
branch (i.e. promoted to baseline) then we can conclude
that despite the fact that the eval log for a vector
may differ from baseline, that this difference is
acceptable so long as the diff marker itself has not
also updated (indicating that the change impacted
more or less than what we previously promoted to baseline).