WIP: Add some more interesting code to the holodeck test runner... ... to keep track of changes in evaluation JSON logfiles produced by executing holodeck test vectors through the runner.
Use of `git diff` and `git diff-tree` tools I think will
allow us to substantially lock down expected result JSON
even in cases where there's transitive (volatile) values
present in the harness evaluation log JSON. This is because
(I think) we can deduce higher-order patterns from git-diff-tree
that allow us to accept changes in a specific file at specific
locations to be accepted as volatile (and thus excluded from
the comparison for the purposes of grossly determining if
a specific test vector evaluation actually did anything that
we haven't previously seen before. It may not work out to be
all that simple to automate entirely. But, even so having
these hashes around provides a lot of extra information that's
very useful for assessing the fitness of the holodeck runner
and harnesses. And, for inspecting rather deeply and at low
cost the results of punching test vectors through harnesses
that box your modules(s)-under-test.