WIP: Update the expected baseline change lines (hunk size 0) for test vector yHBXQnMzTc-uD67HICQx that targets non-idempotent test harness TLUZ3YPUTXK8fXhh6t3-Ew (harness test #5).
To understand this look at the test vector eval log:

`./TESTS/logs/eval/yHBXQnMzTc-uD67HICQx.json`

The test harness will generate a new block of random IRUTs every time it executes. These manifest as ten-lines in the test vector evaluation JSON that change every single time we execute the yHBXQnMzTc-uD67HICQx vector.

`./TESTS/logs/eval/yHBXQnMzTc-uD67HICQx-git-diff`

... shows the difference between the last committed version of `yHBXQnMzTc-uD67HICQx.json` and the currently modified `yHBXQnMzTc-uD67HICQx.json` created by executing the holodeck runner.

Note that it too is volatile (makes sense --- it's the difference of two versions of `yHBXQnMzTc-uD67HICQx.json` that each contain ten lines with different IRUTs).

`./TESTS/logs/eval/yHBXQnMzTc-uD67HICQx-change-lines`

... Does not contain the actual difference data (i.e. the patch details are filtered out). It contains only the diff chunk markers demarkating the region(s) of `TESTS/logs/eval/yHBXQnMzTc-uD67HICQx.json` that have been modified (without regard for the values changed this represents just where values have changed)

And, these are typically stable (I assert) unless and until the test vectors are re-executed against an updated implementation that changes the details of where data changes (this will change the change-lines file which is red flag).

This is why it's important to work to:

- minimize use of non-idempotent functions (idempotent functions are money in the bank once you have test coverage).
- when it's necessary to build non-idempotent functions, endeavor to use the lowest possible git diff hunk size possible to achieve a stable change lines output. You don't want false failures. But, you also want to be alerted to basically any change in the underlying behavior and be given a change to inspect. So, start at hunk size 0 and work up.