WIP: Push all the current CellModel exports through the CellModel holodeck harness. This provides really detailed coverage. Actually, an overwhelming ammount of information. Currently, @encapsule/ holistic-app-client-sml exports a CellModel that produces a 70MB JSON serialization. This is ridiculous...
I think that we need to define, in particular, that ControllerAction
and TransitionOperator plug-in instances are really not a serializable
thing. If you call toJSON() on an instance you'll get the instance
metadata and some identifiers that are useful for tracking back
information in the SOURCES tree as required. Otherwise, you'll not
get the full core dump. This will reduce a lot of noise.

For APM we can safely (I think serialize the digraph).

And, with the above small changes then CellModel is probably small enough so that we don't have to cut much deeper (at least not now). I can imaging CellModel's getting fucking huge...