TEST RELEASE @encapsule/holistic v0.0.39-hakai-rc5
... calling this good enough for now on converting toolchains
and merging back to vNext work. We'll see how the latest of
dropping a .yarnrc --ignore-optional in a derived service/app
gets us by until @babel v8 comes out and upgrades their dep
on chokidar and fsevents which turns out to be currently
problematic on MacOS. I am unaffected on Linux.