WIP: Stash work-in-progress on holodeck2 (actually making... ... slow progress towards being done with that stupid little mechanism I think).
Going to hop over and fix some bugs in @encapsule/holarchy
and add some additional holodeck v1 coverage to ensure the
fixes. It should be simple enough to convert all hd1 to hd2
in an hour or two once hd2 is done. So, not going to worry
about it now and do what needs to get done w/hd1.