Bump to v0.0.43 gitnadoiks ...
... Plus, update to Makefile to fix bug in env_clean target intended to recursively delete the currently-installed ./node_modules directory.

I am noticing on a fresh install of Debian 10 Buster that I am having trouble building holistic-master due to some as-yet-undiagnosed problem (seemingly) with @babel versions? Works fine on the laptop AFAIK. But, this new VM on the workstation is not working as expected. I think perhaps a @babel update is necessary based on comments online. But, it's rather odd at the moment - can't see any differences (yarn.lock doesn't get updated after removing node_modules and deleting the local yarn mirror cache. So, that's strange.