BUGFIX: Update the mission-critical CellModel::constructor filter
I found a bug in CellModel's constructor building out CellModel's
to test Cell Process Manager's process tree digraph transition
operators. The problem was like this:

CellModel A (declares X, Y, Z as subcell dependencies)
CellModel X (just a CellModel)
CellModel Y (declares X, Z as subcell dependencies)
CellModel Z (declares X as subcell dependency)

A <- X
A <- Y <- X
A <- Y <- Z <- X
Z <- Z <- X

Anyway... So recall that a CellModel's private digraph contains
a vertex for every artifact that the CellModel represents and
that this includes a vertex for itself. However, unlike every
other artifact vertex in a CellModel digraph the vertex that
is allocated for the entity itself (the CellModel) _does not_
have a link to an artifact (i.e. the vertex props does not
have an artifact reference to a CellModel like every other
vertex in the digraph).

This is because if we did that we would no longer be able
to conveniently serialize a CellModel to JSON because this
would introduce a cycle into the serialization data. So,
we do not do that.

However, I also missed this case in the the CellModel constructor's
complex rules for auditing the merge of the CellModel's and this
was causing a very mysterious unhandled exception during specific
test cases of CellModel constructor --- very concerning. But,
ultimately no big deal to address once I understood what was
going on.