RELEASE v0.0.45-trinity (w/outstanding test debt >:/)
We need to move on up the stack in the derived vp5 client
application and so I am releasing v0.0.45-trinity as-it-is
knowing that there may be some bugs that need to be fixed
in the CPM transition operators? No matter what these operators
(all the operators for that matter) must be 100% tested before
we ever try to build application #2 w/@encapsule/holistic platform.