WIP Preliminary udpates to HolisticAppClient class...
... But, actually in order to test the client we need to
onfork existing servers that are now broken as I've started
to take control of metadata (finally) in HolisticAppCommon.

So, back the server; taking a large hammer to vp5 derived
app in order to remove all the cruft and clutter from the
mainline build.

By the time I am done w/this there will be only a small
number of files in the vp5 repo.

And, they can have their demo.

And, we will have this...