WIP on v0.0.49-spectrolite refactoring (fascade pattern to contain the entire developer API for everything...)
Fuck it... It's long overdue and I just do not give a shit
anymore; this has to work perfectly or it's not worth anything.
And, it's worth a lot to me to make sure it works perfectly
for others who will never be able to fully understand or likely
recreate anything like this in their lifetime.