WIP lock-in more incremental work to refactor the HTML5 service kernel boot sequence. This commit specifically removes most (nearly all) of the holistic HTML5 service kernel's previously-defined lifecycle actions. We are now only really calling "start", "error", and "hashroute".
Things are still broken in this build.

About to push all the information required by a derived HTML5 service process to initialize itself given the full range of possibile starting places that we allow (still quite a few but thankfully far fewer than is possible given standards (or lack thereof).