WIP snapshot of a bunch of chaos (mostly in my head)...
Most important thing is the introduction of CellModelArtifactSpace ES6 class export to @encapsule/holarchy RTL. This has been gestating in @encapsule/holarchy-cm for the past several days but is being demoted down to @encapsule/holarchy layer. Note that that it's currently stand-alone - there is nobody using this facility in this build - net even other classes in @encapsule/holarchy. It's primarily so that we _could_ someday perhaps use it inside @encapsule/holarchy. But, more immediately because it provides a means for deducing CellModel artifact IRUT's it's only logical to export it from @encapsule/holarchy RTL.