WIP Fix bug in @encapsule/holarchy CellModel constructor request filter and spec...
The issue was that the CellModel::constructor filter input
spec included a reference to CellModel::constructor function's
this + class reference that the actual constructor function
splices into the request descriptor object it receives prior
to calling the constructor filter. We want to do this splicing
in cm-method-constructor-filter.js and not in the
cm-method-constructor-input-spec.js module so that the spec
accurately reflects what a user of CellModel class needs to
pass into the the actual constructor function.