SNAPSHOT of WIP on new @encapsule/holarchy RTL holodeck tests...
This build includes just a few small changes to CellModelArtifactSpace class implementation + yesterday's effort (still not complete) to provide zero ambiguity about CellModelArtifactSpace behavior.

Changes here were propogated up through dependencies so that the tests seeming are stable. But, there's a ton of missing coverage in @encapsule/holodeck-assets... So, we won't really know until we try it in THE FUCKING APP #1.