... Working on some very low-level changes related to localhost development, and deployment on AWS ElasticBeanstalk.
In a nutshell, the development process used at Viewpath is
not adequate to support the rapid development of UX based on
software models. And, supported only a low-level API for extending
the backend using @Encapsule/arccore.filter and RMDR pattern dispatch.

What we need more generally is for @Encapsule/holistic package to
continue to work as it does not; a repo you clone and locally install
in order to gain access to the "holistic app platform" RTL and tools.

Thereafter, `appgen` (or some suitable replacement) will need to
support the (a) initialization of a default application iff applied to
an empty repo (b) support update of holistic platform RTL's and build
infrastructure as it does today (c) fully support the creation of an
X-service-image npm package that may be installed and subsequently
used to launch an X service instance in a Node.js environment.

This last item (c) is not new; it's the model we used previously when
I was working for Quantcast where we used AWS ElasticBeanstalk.

Current plan is to move everything to AWS EB deployment. And, this means re-learning and sorting out all the various updates since the last time I spent serious time with the AWS SDK's and tools.

Oh, and it seems AWS isn't super keen on yarn package manager. And, so we're moving back to npm. Hopefully, I didn't missing anything. But, first step is relatively non-invasive and involves only getting the final BUILD/runtime-phase3/package.json dependencies setup properly (and figuring out how to make all the @Encapsule/holistic RTL packages available for a deploy-time npm install --production on the X-service-image npm package to be generated by executing the X service repo's appgen-created Makefile to (re)build the X-service-image package.